+++ internal/scheduler/database.go
@@ -1,0 +1,17 @@
+package scheduler
+
+import (
+	"context"
+	"log"
+
+	firestoreDB "cloud.google.com/go/firestore"
+	"github.com/capymind/internal/firestore"
+)
+
+func createDBClient(ctx context.Context) *firestoreDB.Client {
+	var client, err = firestore.NewClient(ctx)
+	if err != nil {
+		log.Printf("[Scheduler] Error creating firestore client, %s", err.Error())
+	}
+	return client
+}
